% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables.R
\name{block2prec_by_county}
\alias{block2prec_by_county}
\title{Aggregate a Block Table by Matches Within County}
\usage{
block2prec_by_county(block_table, precinct, precinct_county_fips, epsg = 3857)
}
\arguments{
\item{block_table}{Block table, usually from \code{\link[=create_block_table]{create_block_table()}}.}

\item{precinct}{\code{sf} object of target precinct shapes.}

\item{precinct_county_fips}{Name of the county identifier column in \code{precinct}.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
dataframe with one row per row of \code{precinct}, ordered to align with
the input precinct object.
}
\description{
Variant of \code{\link[=block2prec]{block2prec()}} that matches blocks to precincts separately within
each county. This helps avoid cross-county mismatches when county boundaries
and census blocks do not line up cleanly.
}
\examples{
\dontrun{
# Need Census API
data(towns)
towns$fips <- '087'
block <- create_block_table('NY', 'Rockland')
block2prec_by_county(block, towns, 'fips')
}
}
\concept{datatable}
