% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{check_contiguity}
\alias{check_contiguity}
\alias{cct}
\alias{ccm}
\title{Check Contiguity by Group}
\usage{
check_contiguity(adj, group)

cct(adj, group)

ccm(adj, group)
}
\arguments{
\item{adj}{adjacency list}

\item{group}{array of group identifiers. Typically district numbers or county names.
Defaults to 1 if no input is provided, checking that the adjacency list itself is
one connected component.}
}
\value{
tibble with contiguity indicators. Each row is the units of \code{adj}. Columns include
\itemize{
\item \code{group} Values of the inputted \code{group} argument. If \code{group} is not specified, then all values
will be 1.
\item \code{component} A number for each contiguous set of units within a \code{group}. If all units within a
\code{group} are contiguous, all values are 1. If there are two sets, each discontiguous with
the other, the larger one will be numbered 1 and the smaller one will be numbered as 2.
}
}
\description{
Identify contiguous sets of units and numbers each set. Can be extended to repeat the procedure
within a subgeography.
}
\examples{
data(checkerboard)
adj <- adjacency(checkerboard)
# These each indicate the graph is connected.
check_contiguity(adj) # all contiguous
# If there are two discontiguous groups, there will be 2 values of `component`
cct(adj)
ccm(adj)

}
\concept{fix}
