% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{check_polygon_contiguity}
\alias{check_polygon_contiguity}
\title{Check Polygon Contiguity}
\usage{
check_polygon_contiguity(shp, group, epsg = 3857)
}
\arguments{
\item{shp}{An \code{sf} dataframe.}

\item{group}{Unquoted column name in \code{shp} giving the grouping variable. Use a
row identifier when checking whether individual rows are multipart.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
tibble in the same format as \code{\link[=check_contiguity]{check_contiguity()}}.
}
\description{
Cast \code{shp} to component polygons, build the adjacency, and check the contiguity.
Avoids issues where a precinct is actually a multipolygon
}
\examples{
data(checkerboard)
check_polygon_contiguity(checkerboard, i)
}
\concept{fix}
