% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{checkerboard}
\alias{checkerboard}
\title{Checkerboard}
\format{
An sf dataframe with 64 observations
}
\usage{
data("checkerboard")
}
\description{
This data set contains 64 squares in an 8x8 grid, like a checkerboard.
}
\examples{
data("checkerboard")
}
\concept{data}
