% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{checkerboard_adj}
\alias{checkerboard_adj}
\title{Checkerboard Adjacency}
\format{
A list with 64 entries
}
\usage{
data("checkerboard_adj")
}
\description{
This data contains a zero indexed adjacency list for the checkerboard dataset.
}
\examples{
data("checkerboard_adj")
}
\concept{data}
