% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heda.R
\name{clean_heda}
\alias{clean_heda}
\title{Clean HEDA Names}
\usage{
clean_heda(data)
}
\arguments{
\item{data}{sf tibble from HEDA}
}
\value{
data with cleaned names
}
\description{
Clean HEDA Names
}
\examples{
# TODO
}
\concept{datasets}
