% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vest.R
\name{clean_vest}
\alias{clean_vest}
\title{Clean VEST Column Names}
\usage{
clean_vest(data)
}
\arguments{
\item{data}{\code{sf} tibble from \code{\link[=get_vest]{get_vest()}}.}
}
\value{
input object with renamed columns
}
\description{
Rename VEST election-result columns into a more uniform
\code{office_year_party_candidate} style.
}
\examples{
data(va18sub)
va <- clean_vest(va18sub)
}
\concept{datasets}
