% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables.R
\name{create_block_group_table}
\alias{create_block_group_table}
\title{Create Block Group Level Data}
\usage{
create_block_group_table(state, county, geography = TRUE, year = 2019)
}
\arguments{
\item{state}{Required. Two letter state postal code.}

\item{county}{Optional. Name of county.  If not provided, returns block groups for the entire state.}

\item{geography}{Defaults to TRUE. Whether to return the geography or not.}

\item{year}{year, must be >= 2009 and <= 2019.}
}
\value{
dataframe with data for each block group in the selected region. Data includes
2 sets of columns for each race or ethnicity category: population (pop), voting age
population (vap)
}
\description{
Create Block Group Level Data
}
\examples{
\dontrun{
# Relies on Census Bureau API
bg <- create_block_group_table('NY', 'Rockland', year = 2018)
}
}
\concept{datatable}
