% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_up}
\alias{estimate_up}
\title{Aggregate Values Up Using Precomputed Matches}
\usage{
estimate_up(value, group)
}
\arguments{
\item{value}{Numeric vector to aggregate. Defaults to \code{1}.}

\item{group}{Integer vector of length \code{length(value)} giving the destination
row for each value, often from \code{\link[=geo_match]{geo_match()}}.}
}
\value{
numeric vector of length \code{max(group)} with values aggregated by group
}
\description{
Non-geographic companion to \code{\link[=geo_estimate_up]{geo_estimate_up()}}. Use this when you already
have a vector of group assignments.
}
\examples{
library(dplyr)
set.seed(1)
data(checkerboard)
counties <- checkerboard |>
  group_by(id <= 32) |>
  summarize(geometry = sf::st_union(geometry)) |>
  mutate(pop = c(100, 200))
matches <- geo_match(checkerboard, counties)
estimate_up(value = checkerboard$i, group = matches)
}
\concept{estimate}
