% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{geo_estimate_up}
\alias{geo_estimate_up}
\title{Aggregate Values Up to a Larger Geography}
\usage{
geo_estimate_up(from, to, value, method = "center", epsg = 3857)
}
\arguments{
\item{from}{Smaller geography level.}

\item{to}{Larger geography level.}

\item{value}{Numeric vector of length \code{nrow(from)} to aggregate. Defaults to
\code{1}.}

\item{method}{Matching method passed to \code{\link[=geo_match]{geo_match()}}.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
numeric vector of length \code{nrow(to)} with values aggregated by group
}
\description{
Aggregate values from a smaller geography to a larger geography after matching
each row of \code{from} to a row of \code{to}. This is commonly used to roll block-level
counts up to precincts or districts.
}
\details{
Groups in \code{to} with no matched rows are included in the output and receive \code{0}.
}
\examples{
library(dplyr)
set.seed(1)
data(checkerboard)
counties <- checkerboard |>
  group_by(id <= 32) |>
  summarize(geometry = sf::st_union(geometry)) |>
  mutate(pop = c(100, 200))
geo_estimate_up(from = checkerboard, to = counties, value = checkerboard$i)
}
\concept{estimate}
