% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{geo_plot}
\alias{geo_plot}
\title{Plot an \code{sf} Object with Row Numbers}
\usage{
geo_plot(shp)
}
\arguments{
\item{shp}{An \code{sf} object.}
}
\value{
\code{ggplot} object
}
\description{
Quick diagnostic plot that labels each row with its row number.
}
\examples{
data(checkerboard)
geo_plot(checkerboard)

}
\concept{plot}
