% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{geo_plot_group}
\alias{geo_plot_group}
\title{Plot Groups with Connected Components Colored}
\usage{
geo_plot_group(shp, adj, group, save = FALSE, path = "")
}
\arguments{
\item{shp}{An \code{sf} object.}

\item{adj}{Adjacency list.}

\item{group}{Optional vector of group identifiers, typically district numbers
or county names.}

\item{save}{Logical. If \code{TRUE}, save each plot to disk.}

\item{path}{Directory prefix used when \code{save = TRUE}.}
}
\value{
list of \code{ggplot} objects, one per unique group
}
\description{
Plot Groups with Connected Components Colored
}
\examples{
library(dplyr)
data('checkerboard')
data('checkerboard_adj')

checkerboard <- checkerboard |> mutate(discont = as.integer(j == 5 | j == 6))

p <- geo_plot_group(checkerboard, checkerboard_adj, checkerboard$discont)

p[[1]]
p[[2]]
}
\concept{plot}
