% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{geo_sort}
\alias{geo_sort}
\title{Sort Features by Northwest-to-Southeast Order}
\usage{
geo_sort(shp, epsg = 3857)
}
\arguments{
\item{shp}{\code{sf} dataframe.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
reordered \code{sf} dataframe
}
\description{
Reorder rows by the centroid distance from the northwest corner of the overall
bounding box.
}
\examples{
data(checkerboard)
geo_sort(checkerboard)
}
\concept{fix}
