\name{geomander-package}
\alias{geomander-package}
\alias{geomander}
\docType{package}
\title{\packageTitle{geomander}}
\description{\packageDescription{geomander}}
\section{Package Content}{\packageIndices{geomander}}
\author{\packageAuthor{geomander}}
\section{Maintainer}{\packageMaintainer{geomander}}
\keyword{package}
