% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{geos_centerish}
\alias{geos_centerish}
\title{Get the kind of center of each shape}
\usage{
geos_centerish(shp, epsg = 3857)
}
\arguments{
\item{shp}{An \code{sf} dataframe.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
geos geometry vector of points.
}
\description{
Returns points within the shape, near the center.
Uses the centroid if that's in the shape, or point on surface if not.
}
\examples{
data(towns)
geos_centerish(towns)

}
\concept{center}
