% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{geos_circle_center}
\alias{geos_circle_center}
\title{Get the centroid of the maximum inscribed circle}
\usage{
geos_circle_center(shp, tolerance = 0.01, epsg = 3857)
}
\arguments{
\item{shp}{An sf dataframe}

\item{tolerance}{positive numeric tolerance to simplify by. Default is 0.01.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
A geos geometry list
}
\description{
Returns the centroid of the largest inscribed circle for each shape
}
\examples{
data(towns)
geos_circle_center(towns)

}
\concept{center}
