% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpvnearme.R
\name{get_rpvnearme}
\alias{get_rpvnearme}
\title{Get the RPV Near Me Dataset}
\usage{
get_rpvnearme(state, version = c(1, 2))
}
\arguments{
\item{state}{State postal code.}

\item{version}{Dataset version. Use \code{1} for the original release or \code{2} for
the extended release.}
}
\value{
tibble of precinct-level estimates of party vote by race
}
\description{
Download precinct-level racially polarized voting estimates from the RPV Near
Me repository.
}
\examples{
get_rpvnearme('DE')
}
