% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vest.R
\name{get_vest}
\alias{get_vest}
\title{Get a Voting and Election Science Team ("VEST") Dataset}
\usage{
get_vest(state, year, path = tempdir(), clean_names = TRUE, epsg = 3857, ...)
}
\arguments{
\item{state}{Two-letter state abbreviation.}

\item{year}{Election year, currently one of \code{2016} through \code{2021}.}

\item{path}{Directory used for extracted files. Defaults to \code{tempdir()}.}

\item{clean_names}{Logical. If \code{TRUE}, rename election columns into a more
consistent scheme.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}

\item{...}{Additional arguments passed to \code{\link[sf:st_read]{sf::read_sf()}}.}
}
\value{
\code{sf} tibble containing the requested VEST dataset
}
\description{
Download a state-year VEST precinct file from Harvard Dataverse.
}
\examples{
\dontrun{
# Requires Dataverse API
shp <- get_vest('CO', 2020)
}
}
\concept{datasets}
