% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{global_morans}
\alias{global_morans}
\title{Compute Global Moran's I}
\usage{
global_morans(shp, adj, wts, spatial_mat, epsg = 3857)
}
\arguments{
\item{shp}{\code{sf} dataframe. Optional if \code{adj} or \code{spatial_mat} is supplied.}

\item{adj}{Zero-indexed adjacency list. Optional if \code{shp} or \code{spatial_mat} is
supplied.}

\item{wts}{Numeric vector of observed values.}

\item{spatial_mat}{Square spatial weights matrix. Optional if \code{shp} or \code{adj}
is supplied.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
named list returned by the underlying C++ implementation, including the
global statistic and its moments.
}
\description{
Computes the Global Moran's I statistic and expectation. Can produce spatial weights
from an adjacency or sf dataframe, in which case the spatial_mat is a contiguity
matrix. Users can also provide a spatial_mat argument directly.
}
\examples{
library(dplyr)
data('checkerboard')
checkerboard <- checkerboard |> mutate(m = as.numeric((id + i) \%\% 2 == 0))
global_morans(shp = checkerboard, wts = checkerboard$m)
}
\concept{spatcorr}
