% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstar.R
\name{gstar_i}
\alias{gstar_i}
\title{Compute Standardized Getis-Ord G*i}
\usage{
gstar_i(shp, adj, wts, spatial_mat, epsg = 3857)
}
\arguments{
\item{shp}{\code{sf} dataframe. Optional if \code{adj} or \code{spatial_mat} is supplied.}

\item{adj}{Zero-indexed adjacency list. Optional if \code{shp} or \code{spatial_mat} is
supplied.}

\item{wts}{Numeric vector of observed values.}

\item{spatial_mat}{Square spatial weights matrix. Optional if \code{shp} or \code{adj}
is supplied.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
numeric vector of standardized G*i scores, one per observation
}
\description{
Returns standardized local Getis-Ord G*i scores using either an \code{sf} object, a
zero-indexed adjacency list, or a spatial weights matrix.
}
\examples{
library(dplyr)
data('checkerboard')
checkerboard <- checkerboard |> mutate(m = as.numeric((id + i) \%\% 2 == 0))
gstar_i(shp = checkerboard, wts = checkerboard$m)
}
\concept{spatcorr}
