% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heda.R
\name{heda_states}
\alias{heda_states}
\title{List Available HEDA States}
\usage{
heda_states()
}
\value{
character vector of lowercase state abbreviations with packaged HEDA
support.
}
\description{
List Available HEDA States
}
\examples{
heda_states()
}
\concept{datasets}
