% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{local_morans}
\alias{local_morans}
\title{Compute Local Moran's I}
\usage{
local_morans(shp, adj, wts, spatial_mat, epsg = 3857)
}
\arguments{
\item{shp}{\code{sf} dataframe. Optional if \code{adj} or \code{spatial_mat} is supplied.}

\item{adj}{Zero-indexed adjacency list. Optional if \code{shp} or \code{spatial_mat} is
supplied.}

\item{wts}{Numeric vector of observed values.}

\item{spatial_mat}{Square spatial weights matrix. Optional if \code{shp} or \code{adj}
is supplied.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
tibble with columns \code{moran}, \code{expectation}, and \code{variance}
}
\description{
Compute Local Moran's I for each observation using either an \code{sf} object, a
zero-indexed adjacency list, or a spatial weights matrix.
}
\examples{
library(dplyr)
data('checkerboard')
checkerboard <- checkerboard |> mutate(m = as.numeric((id + i) \%\% 2 == 0))
local_morans(shp = checkerboard, wts = checkerboard$m)
}
\concept{spatcorr}
