% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nrcsd}
\alias{nrcsd}
\title{nrcsd}
\format{
An sf dataframe with 1 observation
}
\usage{
data('nrcsd')
}
\description{
The data contains the North Rockland Central School District.
}
\examples{
data('nrcsd')
}
\concept{data}
