% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{orange}
\alias{orange}
\title{orange}
\format{
An sf dataframe with 10034 observations
}
\usage{
data("orange")
}
\description{
This data contains the blocks for Orange County NY, with geographies simplified
to allow for better examples.
}
\details{
It can be recreated with:
orange <- create_block_table('NY', 'Orange')
orange <- rmapshaper::ms_simplify(orange, keep_shapes = TRUE)
}
\examples{
data("orange")
}
\concept{data}
