% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{precincts}
\alias{precincts}
\title{precincts}
\format{
An sf dataframe with 278 observations
}
\usage{
data("precincts")
}
\description{
This data contains the election districts (or precincts) for Rockland County
NY, with geographies simplified to allow for better examples.
}
\examples{
data('precincts')
}
\references{
https://www.rocklandgis.com/portal/apps/sites/#/data/datasets/2d91f9db816c48318848ad66eb1a18e9
}
\concept{data}
