% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rockland}
\alias{rockland}
\title{rockland}
\format{
An sf dataframe with 4764 observations
}
\usage{
data("rockland")
}
\description{
This data contains the blocks for Rockland County NY, with geographies simplified
to allow for better examples.
}
\details{
It can be recreated with:
rockland <- create_block_table('NY', 'Rockland')
rockland <- rmapshaper::ms_simplify(rockland, keep_shapes = TRUE)
}
\examples{
data("rockland")
}
\concept{data}
