% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seam.R
\name{seam_adj}
\alias{seam_adj}
\title{Filter Adjacency to the Edges Along a Seam}
\usage{
seam_adj(adj, shp, admin, seam, epsg = 3857)
}
\arguments{
\item{adj}{Zero-indexed adjacency graph.}

\item{shp}{\code{sf} object containing the administrative identifier.}

\item{admin}{Name of the administrative-unit column in \code{shp}.}

\item{seam}{Length-2 vector of administrative-unit values defining the seam.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
adjacency list containing only cross-seam edges
}
\description{
Keep only the adjacency edges that connect the two sides of a selected seam.
}
\examples{
data('rockland')
data('orange')
data('nrcsd')

o_and_r <- rbind(orange, rockland)
o_and_r <- o_and_r |>
  geo_filter(nrcsd) |>
  geo_trim(nrcsd)
adj <- adjacency(o_and_r)

seam_adj(adj, shp = o_and_r, admin = 'county', seam = c('071', '087'))

}
\concept{seam}
