% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seam.R
\name{seam_rip}
\alias{seam_rip}
\title{Remove Edges Across an Administrative Boundary}
\usage{
seam_rip(adj, shp, admin, seam, epsg = 3857)
}
\arguments{
\item{adj}{Zero-indexed adjacency graph.}

\item{shp}{\code{sf} object containing the administrative identifier.}

\item{admin}{Name of the administrative-unit column in \code{shp}.}

\item{seam}{Vector of administrative-unit values to separate.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
adjacency list with all cross-seam edges removed
}
\description{
Remove adjacency connections between two or more administrative units, such as
counties on opposite sides of a seam.
}
\examples{
data('rockland')
data('orange')
data('nrcsd')

o_and_r <- rbind(orange, rockland)
o_and_r <- o_and_r |>
  geo_filter(nrcsd) |>
  geo_trim(nrcsd)
adj <- adjacency(o_and_r)

seam_rip(adj, o_and_r, 'county', c('071', '087'))

}
\concept{seam}
