% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seam.R
\name{seam_sew}
\alias{seam_sew}
\title{Suggest Edges to Sew a Seam}
\usage{
seam_sew(shp, admin, seam, epsg = 3857)
}
\arguments{
\item{shp}{\code{sf} object containing the administrative identifier.}

\item{admin}{Name of the administrative-unit column in \code{shp}.}

\item{seam}{Length-2 vector of administrative-unit values defining the seam.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
tibble with columns \code{v1} and \code{v2}, suitable for use with \code{\link[=add_edge]{add_edge()}}
}
\description{
Build a set of candidate cross-seam edges by constructing Voronoi cells from
representative points and identifying neighboring cells across the seam.
}
\examples{
data('rockland')
data('orange')
data('nrcsd')

o_and_r <- rbind(orange, rockland)
o_and_r <- o_and_r |>
  geo_filter(nrcsd) |>
  geo_trim(nrcsd)
adj <- adjacency(o_and_r)

adds <- seam_sew(o_and_r, 'county', c('071', '087'))
adj <- adj |> add_edge(adds$v1, adds$v2)

}
\concept{seam}
