% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_precinct}
\alias{split_precinct}
\title{Split a Precinct}
\usage{
split_precinct(lower, precinct, split_by, lower_wt, split_by_id, epsg = 3857)
}
\arguments{
\item{lower}{Lower geography that makes up the precinct, often blocks.}

\item{precinct}{Single-row \code{sf} object giving the precinct to split.}

\item{split_by}{Geography that defines the pieces to split the precinct into.}

\item{lower_wt}{Optional numeric vector of length \code{nrow(lower)} to aggregate
within the split pieces, such as population or VAP.}

\item{split_by_id}{Optional column name in \code{split_by} to copy onto the output.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
\code{sf} dataframe with one row per observed split piece
}
\description{
States often split a precinct when they create districts but rarely provide the
geography for the split precinct. This allows you to split a precinct using a
lower geography, typically blocks.
}
\details{
The function first filters \code{lower} and \code{split_by} to the selected precinct,
matches lower-level units to \code{split_by}, and unions matched pieces. When
\code{lower_wt} is supplied, a \code{wt} column is added with summed weights.
}
\examples{
library(sf)
data(checkerboard)
low <- checkerboard |> dplyr::slice(1:3, 9:11)
prec <- checkerboard |>
  dplyr::slice(1:3) |>
  dplyr::summarize(geometry = sf::st_union(geometry))
dists <- checkerboard |>
  dplyr::slice(1:3, 9:11) |>
  dplyr::mutate(dist = c(1, 2, 2, 1, 3, 3)) |>
  dplyr::group_by(dist) |>
  dplyr::summarize(geometry = sf::st_union(geometry))

split_precinct(low, prec, dists, split_by_id = 'dist')
}
\concept{fix}
