% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{suggest_neighbors}
\alias{suggest_neighbors}
\title{Suggest Neighbors for Lonely Precincts}
\usage{
suggest_neighbors(shp, adj, idx, neighbors = 1)
}
\arguments{
\item{shp}{\code{sf} object used to compute representative points.}

\item{adj}{Adjacency list.}

\item{idx}{Optional integer vector of row indices to repair. If omitted, the
function uses rows with no neighbors.}

\item{neighbors}{Number of candidate neighbors to return for each row in
\code{idx}.}
}
\value{
tibble with columns \code{x} and \code{y}, giving pairs of row indices that
could be connected with \code{\link[=add_edge]{add_edge()}}.
}
\description{
For precincts which have no adjacent precincts, this suggests the nearest precinct
as a friend to add. This is useful for when a small number of precincts are disconnected
from the remainder of the geography, such as an island.
}
\examples{
library(dplyr)
data(va18sub)
va18sub <- va18sub |> filter(!VTDST \%in\% c('000516', '000510', '000505', '000518'))
adj <- adjacency(va18sub)
suggests <- suggest_neighbors(va18sub, adj)
adj <- adj |> add_edge(v1 = suggests$x, v2 = suggests$y)

}
\concept{fix}
