% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{towns}
\alias{towns}
\title{towns}
\format{
An sf dataframe with 7 observations
}
\usage{
data("towns")
}
\description{
This data contains 7 town boundaries for the towns which overlap
North Rockland School District in NY.
}
\examples{
data("towns")
}
\references{
https://www.rocklandgis.com/portal/apps/sites/#/data/items/746ec7870a0b4f46b168e07369e79a27
}
\concept{data}
