% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{va18sub}
\alias{va18sub}
\title{va18sub}
\format{
An sf dataframe with 116 observations
}
\usage{
data("va18sub")
}
\description{
This data contains a 116 precinct subset of Virginia from the 2018 Senate race.
Constains results for counties Henrico (087) and Henry (089).
}
\examples{
 \dontrun{data("va18sub")}
}
\references{
Voting and Election Science Team, 2019, "va_2018.zip", 2
018 Precinct-Level Election Results, 
https://doi.org/10.7910/DVN/UBKYRU/FQDLOO, Harvard Dataverse, V4
}
