% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{va_blocks}
\alias{va_blocks}
\title{va_blocks}
\format{
An sf dataframe with 6354 observations
}
\usage{
data("va_blocks")
}
\description{
This data contains the blocks Henrico County, VA with geographies simplified
to allow for better examples.
}
\details{
blocks87 <- create_block_table(state = 'VA', county = '087')
va_blocks <- rmapshaper::ms_simplify(va_blocks, keep_shapes = TRUE)
}
\examples{
data('va_blocks')
}
\concept{data}
