% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{va_vtd}
\alias{va_vtd}
\title{va_vtd}
\format{
An sf dataframe with 93 observations
}
\usage{
data("va_blocks")
}
\description{
This data contains the blocks for Henrico County, VA with geographies simplified
to allow for better examples.
}
\details{
va_vtd <- tigris::voting_districts(state = 'VA') %>% filter(COUNTYFP10 == '087')
va_vtd <- rmapshaper::ms_simplify(va_vtd, keep_shapes = TRUE)
}
\examples{
data("va_blocks")
}
\concept{data}
