% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vest.R
\name{vest_states}
\alias{vest_states}
\title{List Available VEST States}
\usage{
vest_states(year)
}
\arguments{
\item{year}{Year to inspect in the VEST Dataverse collection.}
}
\value{
character vector of state abbreviations available for that year
}
\description{
List Available VEST States
}
\examples{
\dontrun{
# Requires Dataverse API
vest_states(2020)
}
}
\concept{datasets}
