% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency}
\alias{adjacency}
\title{Build Adjacency List}
\usage{
adjacency(shp, zero = TRUE, rook = TRUE)
}
\arguments{
\item{shp}{sf dataframe}

\item{zero}{Default is TRUE for zero indexed. FALSE gives a one indexed adjacency list.}

\item{rook}{Default is TRUE for rook adjacency. FALSE gives queen adjacency}
}
\value{
list with nrow(shp) entries
}
\description{
This is similar to the old version of redist's adjacency function which has been
replaced by sf internals. This is faster but less reliable. This wraps spdep::poly2nb()
and replaces the indicator for no adjacent precincts with a integer(0)
}
\examples{
data(precincts)
adj <- adjacency(precincts)
}
\concept{fix}
