% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alarm.R
\name{alarm_states}
\alias{alarm_states}
\title{List Available States from ALARM Data}
\usage{
alarm_states()
}
\value{
character abbreviations for states
}
\description{
List Available States from ALARM Data
}
\examples{
\donttest{
# relies on internet availability
alarm_states()
}
}
