% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{check_contiguity}
\alias{check_contiguity}
\title{Check Contiguity by Group}
\usage{
check_contiguity(adjacency, group)
}
\arguments{
\item{adjacency}{adjacency list}

\item{group}{array of group identifiers. Typically district numbers or county names.}
}
\value{
tibble with a column for each of inputted group, created group number, and the 
identified connected component number
}
\description{
Check Contiguity by Group
}
\examples{
data(checkerboard)
adj <- adjacency(checkerboard)
check_contiguity(adj)

}
\concept{fix}
