% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{check_polygon_contiguity}
\alias{check_polygon_contiguity}
\title{Check Polygon Contiguity}
\usage{
check_polygon_contiguity(shp, group)
}
\arguments{
\item{shp}{An sf data frame}

\item{group}{unqouted name of group identifier in shp}
}
\value{
tibble with a column for each of inputted group, created group number, and the 
identified connected component number
}
\description{
Cast `shp` to component polygons, build the adjacency, and check the contiguity.
Avoids issues where a precinct is actually a multipolygon
}
\examples{
data(checkerboard)
check_contiguity(checkerboard, i)

}
\concept{fix}
