% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vest.R
\name{clean_vest}
\alias{clean_vest}
\title{Clean VEST Names}
\usage{
clean_vest(data)
}
\arguments{
\item{data}{sf tibble from VEST}
}
\value{
data with cleaned names
}
\description{
Clean VEST Names
}
\examples{
data(va18sub)
va <- clean_vest(va18sub)
}
\concept{datasets}
