% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_connections.R
\name{count_connections}
\alias{count_connections}
\title{Count Times Precincts are Connected}
\usage{
count_connections(dm, normalize = FALSE)
}
\arguments{
\item{dm}{district membership matrix}

\item{normalize}{Whether to normalize all values by the number of columns.}
}
\value{
matrix with the number of connections between precincts
}
\description{
Count Times Precincts are Connected
}
\examples{
set.seed(1)
dm <- matrix(sample(1:2, size = 100, TRUE), 10)
count_connections(dm)
}
\concept{leftover}
