% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables.R
\name{create_block_table}
\alias{create_block_table}
\title{Create Block Level Data}
\usage{
create_block_table(state, county, geography = TRUE, year = 2010)
}
\arguments{
\item{state}{Required. Two letter state postal code.}

\item{county}{Optional. Name of county.  If not provided, returns blocks for the entire state.}

\item{geography}{Defaults to TRUE. Whether to return the geography or not.}

\item{year}{year, must be 2010 at the moment. 2020 to be added once available. 2000 if rereleased.}
}
\value{
dataframe with data for each block in the selected region
}
\description{
Creates a block level dataset, using the decennial census information, with the 
standard redistricting variables.
}
