% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{geo_estimate_down}
\alias{geo_estimate_down}
\title{Estimate Down Geography Levels}
\usage{
geo_estimate_down(from, to, wts, value, method = "center")
}
\arguments{
\item{from}{Larger geography level}

\item{to}{smaller geography level}

\item{wts}{numeric vector of length nrow(to). Defaults to 1. Typically population or VAP, as a weight to give each precinct.}

\item{value}{numeric vector of length nrow(from). Defaults to 1. Typically electoral outcomes, as a value to estimate down into blocks.}

\item{method}{string from center, centroid, point, or area for matching levels}
}
\value{
numeric vector with each value split by weight
}
\description{
Simple method for estimating data down to a lower level. This is most often useful
for getting election data down from a precinct level to a block level in the case
that a state or other jurisdiction split precincts when creating districts. Geographic
parter to estimate_down.
}
\examples{
\dontrun{
data("va18sub")
va18sub <- va18sub \%>\% filter(COUNTYFP == '087')
block <- create_block_table(state = 'VA', county = '087')  
disagg <- geo_estimate_down(from = va18sub, to = block, wts = block$VAP, value = va18sub$G18USSRSTE)
}
}
