% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{geo_estimate_up}
\alias{geo_estimate_up}
\title{Estimate Up Geography Levels}
\usage{
geo_estimate_up(from, to, value, method = "center")
}
\arguments{
\item{from}{smaller geography level}

\item{to}{larger geography level}

\item{value}{numeric vector of length nrow(from). Defaults to 1.}

\item{method}{string from center, centroid, point, or area for matching levels}
}
\value{
numeric vector with each value aggregated by grou
}
\description{
Simple method for aggregating data up to a higher level This is most often useful
for getting population data from a block level up to a precinct level.
Geographic parter to estimate_up.
}
\examples{
\dontrun{
data("va18sub")
va18sub <- va18sub \%>\% filter(COUNTYFP == '087')
block <- create_block_table(state = 'VA', county = '087')  
agg <- geo_estimate_down(from = block, to = va18shp, value = block$Total)
}
}
