% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{geo_plot}
\alias{geo_plot}
\title{Plots a Shape with Row Numbers as Text}
\usage{
geo_plot(shp)
}
\arguments{
\item{shp}{An sf shapefile}
}
\value{
ggplot
}
\description{
One liner to plot a shape with row numbers
}
\examples{
data(checkerboard)
geo_plot(checkerboard)

}
\concept{plot}
