% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{geo_sort}
\alias{geo_sort}
\title{Sort Precincts}
\usage{
geo_sort(shp)
}
\arguments{
\item{shp}{sf dataframe, required.}
}
\value{
sf dataframe
}
\description{
Reorders precincts by distance from the NW corner of the bounding box.
}
\examples{
data(checkerboard)
geo_sort(checkerboard)

}
\concept{fix}
