% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alarm.R
\name{get_alarm}
\alias{get_alarm}
\title{Get ALARM Dataset}
\usage{
get_alarm(state, geometry = TRUE, file = tempfile(fileext = ".csv"))
}
\arguments{
\item{state}{two letter state abbreviation}

\item{geometry}{Default is TRUE. Add geomeetry to the data?}

\item{file}{file path to save ALARM csv}
}
\value{
tibble with election data and optional geometry
}
\description{
Get ALARM Dataset
}
\examples{
ak <- get_alarm('AK')
}
