% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dra_vtd.R
\name{get_dra}
\alias{get_dra}
\title{Get Dave's Redistricting App Dataset}
\usage{
get_dra(state, year = 2020, geometry = TRUE, epsg = 3857)
}
\arguments{
\item{state}{two letter state abbreviation}

\item{year}{year to get data for. Either \code{2020} or \code{2010}}

\item{geometry}{Default is TRUE. Add geometry to the data?}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}
}
\value{
tibble with election data and optional geometry
}
\description{
Gets a dataset from Dave's Redistricting App.
}
\details{
See the full available data at \url{https://github.com/dra2020/vtd_data}.
}
\examples{
\donttest{
# Takes a few seconds to run
ak <- get_dra('AK')
}
}
\concept{datasets}
