% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpvnearme.R
\name{get_rpvnearme}
\alias{get_rpvnearme}
\title{Get Racially Polarized Voting Dataset from RPV Near Me}
\usage{
get_rpvnearme(state, version = c(1, 2))
}
\arguments{
\item{state}{the state postal code of the state}

\item{version}{the version of the data to use. \code{1} for the original, \code{2} for the extended.}
}
\value{
a tibble of precinct-level estimates of votes (party) by race
}
\description{
Get Racially Polarized Voting Dataset from RPV Near Me
}
\examples{
get_rpvnearme('DE')
}
