% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vest.R
\name{get_vest}
\alias{get_vest}
\title{Get VEST Dataset}
\usage{
get_vest(state, year, path = tempdir(), clean_names = TRUE)
}
\arguments{
\item{state}{two letter state abbreviation}

\item{year}{year in 2016, 2018, or 2020}

\item{path}{folder to put shape in. Default is \code{tempdir()}}

\item{clean_names}{Clean names. Default is \code{TRUE}. If \code{FALSE},
returns default names.}
}
\value{
sf tibble
}
\description{
Get VEST Dataset
}
\examples{
\dontrun{
# Requires Dataverse API
shp <- get_vest('CO', 2020)
}
}
