% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vest.R
\name{get_vest}
\alias{get_vest}
\title{Get Voting and Election Science Team ("VEST") Dataset}
\usage{
get_vest(state, year, path = tempdir(), clean_names = TRUE, epsg = 3857, ...)
}
\arguments{
\item{state}{two letter state abbreviation}

\item{year}{year any in 2016-2021}

\item{path}{folder to put shape in. Default is \code{tempdir()}}

\item{clean_names}{Clean names. Default is \code{TRUE}. If \code{FALSE},
returns default names.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}

\item{...}{additional arguments passed to \code{\link[sf:st_read]{sf::read_sf()}}}
}
\value{
sf tibble
}
\description{
Get Voting and Election Science Team ("VEST") Dataset
}
\examples{
\dontrun{
# Requires Dataverse API
shp <- get_vest('CO', 2020)
}
}
\concept{datasets}
