% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heda.R
\name{heda_states}
\alias{heda_states}
\title{List Available States from HEDA Dataverse}
\usage{
heda_states()
}
\value{
character abbreviations for states
}
\description{
List Available States from HEDA Dataverse
}
\examples{
heda_states()
}
\concept{datasets}
