% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dra.R
\name{r2dra}
\alias{r2dra}
\title{R to DRA}
\usage{
r2dra(precincts, plan, state, path)
}
\arguments{
\item{precincts}{Required. an sf dataframe of precinct shapes}

\item{plan}{Required. Either a vector of district assignments 
or the name of a column in precincts with district assignments.}

\item{state}{Required. the state postal code of the state}

\item{path}{Optional. A path to try to save to. Warns if saving failed.}
}
\value{
tibble with columns Id, as used by DRA, identical to GEOID in census terms and District.
}
\description{
R to DRA
}
\examples{
\dontrun{

}
}
