% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{st_centerish}
\alias{st_centerish}
\title{Get the kind of center of each shape}
\usage{
st_centerish(shp)
}
\arguments{
\item{shp}{An sf shapefile}
}
\value{

}
\description{
Returns points within the shape, near the center.
Uses the centroid if that's in the shape, or point on surface if not.
}
