% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{suggest_component_connection}
\alias{suggest_component_connection}
\title{Suggest Connections for Disconnected Groups}
\usage{
suggest_component_connection(shp, adjacency, group)
}
\arguments{
\item{shp}{}

\item{adjacency}{}

\item{group}{}
}
\value{

}
\description{
Suggests nearest neighbors for connected a disconnected group.
}
