% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{suggest_component_connection}
\alias{suggest_component_connection}
\title{Suggest Connections for Disconnected Groups}
\usage{
suggest_component_connection(shp, adjacency, group)
}
\arguments{
\item{shp}{An sf shapefile}

\item{adjacency}{adjacency list}

\item{group}{array of group identifiers. Typically district numbers or county names.}
}
\value{
tibble with two columns of suggested rows of shp to connect in adj
}
\description{
Suggests nearest neighbors for connected a disconnected group.
}
