% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vest.R
\name{vest_states}
\alias{vest_states}
\title{List Available States from VEST Dataverse}
\usage{
vest_states(year)
}
\arguments{
\item{year}{year in 2016, 2018, or 2020}
}
\value{
character abbreviations for states
}
\description{
List Available States from VEST Dataverse
}
\examples{
\dontrun{
# Requires Dataverse API
vest_states(2020)
}
}
\concept{datasets}
